# Hash five bytes instead of six